Moved beamWeapon to own file
