+++ CMakeLists.txt
@@ -131,6 +131,7 @@ set(SOURCES
 	src/gui/gui2_overlay.cpp
 	src/spaceObjects/spaceStation.cpp
 	src/spaceObjects/spaceship.cpp
+	src/spaceObjects/beamWeapon.cpp
 	src/spaceObjects/wormHole.cpp
 	src/spaceObjects/spaceObject.cpp
 	src/spaceObjects/nebula.cpp
