+++ src/spaceObjects/beamWeapon.cpp
@@ -1,0 +1,6 @@
+#include "beamWeapon.h"
+
+BeamWeapon::BeamWeapon()
+{
+
+}
